break(ws): move `WebSocketPair` global to `worktop/cfw` module
